requirements.txt
# Telegram Bot
python-telegram-bot[all]==20.7

# HTTP Client
httpx==0.26.0

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-mock==3.12.0
